moved get_grid_coordinates from utils.py to grid.py removed unused import in utils.py added new function: get_empty_grid. It copys the coordinates grid and fills it with empty values
